refactor: define table header color globally
