+++ gravitee-apim-portal-webui-next/src/components/subscription/webhook/consumer-configuration-retry/index.ts
@@ -13,6 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './webhook/consumer-configuration';
-export * from './webhook/consumer-configuration-headers';
-export * from './webhook/consumer-configuration-retry';
+export * from './consumer-configuration-retry.component';
